+++ srcs/nest/src/user/friend/friend.service.ts
@@ -20,13 +20,11 @@ export class FriendService {
         const query: Friend = await this.friendRepository.findOne({
             where: [{ requestUserId: userId, targetUserId: friendId }],
         });
-        // this.logger.debug(`query1 : ${query}`);
         if (query) return query;
 
         const query2: Friend = await this.friendRepository.findOne({
             where: [{ requestUserId: friendId, targetUserId: userId }],
         });
-        // this.logger.debug(`query2: ${query2}`);
         if (!query2) return null; // No matching friend record found
         return query2;
     }
@@ -68,7 +66,7 @@ export class FriendService {
                 if (userId === friend.requestUserId)
                     friendData = await this.userService.findUserById(friend.targetUserId);
                 else friendData = await this.userService.findUserById(friend.requestUserId);
-
+                if (friendData === undefined) continue;
                 friendList.push({ userName: friendData.userName, slackId: friendData.slackId });
             }
         } catch (error) {
@@ -80,13 +78,13 @@ export class FriendService {
 
     async getFriendStatus(userId: number, friendSlackId: string): Promise<FriendStatus> {
         try {
-            this.logger.debug(friendSlackId);
-            const friendId: number = (await this.userService.getUserBySlackId(friendSlackId)).id;
-            const data: Friend = await this.getFriendData(userId, friendId);
-            this.logger.debug(data);
+            const friend: User = await this.userService.getUserBySlackId(friendSlackId);
+            if (friend === undefined) throw new BadRequestException('friendService >> getFriendStatus');
+
+            const data: Friend = await this.getFriendData(userId, friend.id);
             let friendStatus: FriendStatus = FriendStatus.FRIEND;
             if (!data) friendStatus = FriendStatus.UNKNOWN;
-            else if (data.requestUserId === friendId && data.status === FriendStatus.REQUESTED)
+            else if (data.requestUserId === friend.id && data.status === FriendStatus.REQUESTED)
                 friendStatus = FriendStatus.LAGGING;
             else if (data.requestUserId === userId && data.status === FriendStatus.REQUESTED)
                 friendStatus = FriendStatus.REQUESTED;
@@ -101,8 +99,8 @@ export class FriendService {
     async requestFriend(userId: number, friendSlackId: string): Promise<void> {
         try {
             //본인인지 확인
-            const friendId: number = (await this.userService.getUserBySlackId(friendSlackId)).id;
-            if (userId === friendId) return;
+            const friend: User = await this.userService.getUserBySlackId(friendSlackId);
+            if (friend === undefined || userId === friend.id) return;
             //친구 상태 확인
             const status = await this.getFriendStatus(userId, friendSlackId);
             switch (status) {
@@ -116,10 +114,9 @@ export class FriendService {
                     return;
                 //기록 없을 경우 DB처리 진행
                 case FriendStatus.UNKNOWN:
-                    const friendId = (await this.userService.getUserBySlackId(friendSlackId)).id;
                     const newData = this.friendRepository.create({
                         requestUserId: userId,
-                        targetUserId: friendId,
+                        targetUserId: friend.id,
                         status: FriendStatus.REQUESTED,
                     } as Friend);
                     await this.friendRepository.save(newData);
@@ -134,10 +131,10 @@ export class FriendService {
     }
 
     async deleteFriend(userId: number, slackId: string) {
-        //TODO : socketUserService의 friendList 업데이트
         try {
-            const friendId = (await this.userService.getUserBySlackId(slackId)).id;
-            const data = await this.getFriendData(userId, friendId);
+            const friend: User = await this.userService.getUserBySlackId(slackId);
+            if (friend === undefined) return;
+            const data = await this.getFriendData(userId, friend.id);
             if (!data || data.status !== FriendStatus.FRIEND) return;
             await this.friendRepository.delete(data.id);
         } catch (error) {
@@ -159,7 +156,7 @@ export class FriendService {
             if (!requestUsers) return [];
             for (const requestUser of requestUsers) {
                 const friend: User = await this.userService.findUserById(requestUser.friend_requestUserId);
-                invitations.push({ userName: friend.userName, slackId: friend.slackId });
+                if (friend) invitations.push({ userName: friend.userName, slackId: friend.slackId });
             }
             return invitations;
         } catch (error) {
@@ -174,9 +171,10 @@ export class FriendService {
             const status: FriendStatus = await this.getFriendStatus(userId, friendSlackId);
             if (status !== FriendStatus.LAGGING) return;
 
-            const friendId = (await this.userService.getUserBySlackId(friendSlackId)).id;
+            const friend: User = await this.userService.getUserBySlackId(friendSlackId);
+            if (friend === undefined) return;
             await this.friendRepository.update(
-                { requestUserId: friendId, targetUserId: userId },
+                { requestUserId: friend.id, targetUserId: userId },
                 { status: FriendStatus.FRIEND },
             );
         } catch (error) {
@@ -188,10 +186,10 @@ export class FriendService {
     async declineRequest(userId: number, friendSlackId: string) {
         try {
             const status: FriendStatus = await this.getFriendStatus(userId, friendSlackId);
-            if (status !== FriendStatus.LAGGING) return;
+            const friend: User = await this.userService.getUserBySlackId(friendSlackId);
+            if (status !== FriendStatus.LAGGING || friend === undefined) return;
 
-            const friendId = (await this.userService.getUserBySlackId(friendSlackId)).id;
-            const data = await this.getFriendData(userId, friendId);
+            const data = await this.getFriendData(userId, friend.id);
             await this.friendRepository.delete(data.id);
         } catch (error) {
             console.log(error);
